# docs/contracts/1.6/06_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Smart Wardrobe Prototype API
  version: 1.6.0
  description: HLD v1.1 / FSD v1.1 / Backend LLD v1.1 기반. 서버 소유의 REST 계약; Decart·ThinQ
    CLO 외부 실제 API 규격을 주장하지 않음. Bearer 인증과 demo_mode 세션은 프로토타입 정책이며 운영 보안 정책이 아님.
servers:
- url: http://localhost:8000
  description: Local FastAPI prototype
...
security:
- BearerAuth: []
paths:
...
  /api/v1/storage-optimization-jobs:
    post:
      tags:
      - Storage
      summary: 착용 패턴·공간 환경 보관 최적화
      operationId: post_api_v1_storage_optimization_jobs
      x-fsd-ids:
      - STO-002
      - STO-003
      description: mode=WEAR_PATTERN은 STO-002, SPACE_ENVIRONMENT는 STO-003.
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageOptimizeRequest'
      responses:
        '202':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJob'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/storage-actions:
    get:
      tags:
      - Storage
      summary: 보관 제안 조회
      operationId: get_api_v1_storage_actions
      x-fsd-ids:
      - STO-004
      parameters:
      - name: household_id
        in: query
        required: false
        schema: *id002
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PROPOSED
          - APPROVED
          - REJECTED
          - IN_PROGRESS
          - AWAITING_CONFIRMATION
          - COMPLETED
          - FAILED
          - EXPIRED
          - CANCELLED
      - *id004
      - *id005
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageActionList'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/storage-actions/{action_id}:
    get:
      tags:
      - Storage
      summary: 보관 작업 상세
      operationId: get_api_v1_storage_actions_action_id
      x-fsd-ids:
      - STO-004
      parameters:
      - name: action_id
        in: path
        required: true
        schema: *id002
      responses:
        '200':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageAction'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/storage-actions/{action_id}/decision:
    post:
      tags:
      - Storage
      summary: 보관 제안 승인·거절
      operationId: post_api_v1_storage_actions_action_id_decision
      x-fsd-ids:
      - STO-004
      parameters:
      - name: action_id
        in: path
        required: true
        schema: *id002
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageDecisionRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageAction'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/storage-actions/{action_id}/confirm:
    post:
      tags:
      - Storage
      summary: 실제 이동 확인 및 상태 갱신
      operationId: post_api_v1_storage_actions_action_id_confirm
      x-fsd-ids:
      - STO-004
      parameters:
      - name: action_id
        in: path
        required: true
        schema: *id002
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: 재시도 시 동일 명령 결과 반환
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageConfirmRequest'
      responses:
        '201':
          description: 성공 또는 작업 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageAction'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 접근 거부
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 리소스 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 상태 또는 버전 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 외부 서비스 사용 불가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 서버가 발급한 사용자 범위 세션 토큰. 요청 member_id/household_id를 신뢰하지 않고 인증 주체와
        비교.
  schemas:
    Error:
      type: object
      properties:
        code: *id003
        message: *id003
        request_id: *id003
        details:
          type: object
          additionalProperties: true
      required:
      - code
      - message
      - request_id
      additionalProperties: false
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/Error'
      required:
      - error
      additionalProperties: false
    Page:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required:
      - limit
      - offset
      - total
      additionalProperties: false
    AcceptedJob:
      additionalProperties: false
      properties:
        job_id:
          format: uuid
          type: string
        status:
          enum:
          - QUEUED
          - RUNNING
          - SUCCEEDED
          - READY
          - FAILED
          - TIMED_OUT
          - CANCELLED
          type: string
        status_url:
          type: string
      required:
      - job_id
      - status
      - status_url
      type: object
...
    StorageOptimizeRequest:
      additionalProperties: false
      properties:
        household_id:
          format: uuid
          type: string
        member_id:
          format: uuid
          type: string
        mode:
          enum:
          - WEAR_PATTERN
          - SPACE_ENVIRONMENT
          type: string
        dry_run:
          type: boolean
        season:
          enum:
          - SPRING
          - SUMMER
          - AUTUMN
          - WINTER
          - ALL
          type: string
        analysis_window_days:
          maximum: 365
          minimum: 1
          type: integer
        garment_ids:
          items:
            format: uuid
            type: string
          maxItems: 100
          type: array
      required:
      - household_id
      - member_id
      - mode
      type: object
    StorageActionItem:
      additionalProperties: false
      properties:
        id:
          format: uuid
          type: string
        garment_id:
          format: uuid
          type: string
        source_location_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
        destination_location_id:
          format: uuid
          type: string
        status:
          enum:
          - PENDING
          - IN_PROGRESS
          - AWAITING_CONFIRMATION
          - COMPLETED
          - FAILED
          - SKIPPED
          type: string
        expected_garment_version:
          anyOf:
          - type: integer
          - type: 'null'
        expected_state_version:
          anyOf:
          - type: integer
          - type: 'null'
        score:
          anyOf:
          - type: number
          - type: 'null'
        reasons:
          items:
            type: string
          type: array
        confirmed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        confirmation_method:
          anyOf:
          - type: string
          - type: 'null'
        failure_reason:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - id
      - garment_id
      - destination_location_id
      - status
      type: object
    StorageAction:
      additionalProperties: false
      properties:
        id:
          format: uuid
          type: string
        household_id:
          format: uuid
          type: string
        action_type:
          type: string
        status:
          enum:
          - PROPOSED
          - APPROVED
          - REJECTED
          - IN_PROGRESS
          - AWAITING_CONFIRMATION
          - COMPLETED
          - FAILED
          - EXPIRED
          - CANCELLED
          type: string
        reasoning:
          items:
            type: string
          type: array
        items:
          items:
            $ref: '#/components/schemas/StorageActionItem'
          type: array
        version:
          type: integer
        requested_by:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
        approved_by:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
        created_at:
          format: date-time
          type: string
        approved_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        expires_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        rules_version:
          anyOf:
          - type: string
          - type: 'null'
        executable:
          type: boolean
      required:
      - id
      - household_id
      - action_type
      - status
      - reasoning
      - items
      - version
      - created_at
      type: object
    StorageDecisionRequest:
      additionalProperties: false
      properties:
        expected_version:
          minimum: 1
          type: integer
        decision:
          enum:
          - APPROVE
          - REJECT
          - CANCEL
          type: string
      required:
      - expected_version
      - decision
      type: object
    StorageConfirmRequest:
      additionalProperties: false
      properties:
        expected_version:
          minimum: 1
          type: integer
        items:
          items:
            $ref: '#/components/schemas/StorageItemConfirmation'
          maxItems: 100
          minItems: 1
          type: array
      required:
      - expected_version
      - items
      type: object
...
    StorageActionList:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/StorageAction'
          type: array
        page:
          $ref: '#/components/schemas/Page'
      required:
      - items
      - page
      type: object
...
    StorageItemConfirmation:
      additionalProperties: false
      properties:
        garment_id:
          format: uuid
          type: string
        confirmed:
          type: boolean
        observed_location_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
        observed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
        confirmation_method:
          enum:
          - USER
          - SENSOR_VERIFIED
          type: string
        failure_reason:
          anyOf:
          - maxLength: 500
            type: string
          - type: 'null'
      required:
      - garment_id
      - confirmed
      type: object
...